refactor: component package 이동
