+++ android/app/src/main/java/io/coursepick/coursepick/presentation/customcourse/component/CourseDistanceChip.kt
@@ -1,35 +1,31 @@
-package io.coursepick.coursepick.presentation.component
+package io.coursepick.coursepick.presentation.customcourse.component
 
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.colorResource
-import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import io.coursepick.coursepick.R
-import io.coursepick.coursepick.domain.course.Length
+import io.coursepick.coursepick.domain.course.Distance
 import io.coursepick.coursepick.presentation.ui.formattedMeter
 
 @Composable
-fun CourseLengthChip(
-    length: Length,
+fun CourseDistanceChip(
+    distance: Distance,
     modifier: Modifier = Modifier,
 ) {
-    Row(
+    Box(
         modifier =
             modifier
                 .height(28.dp)
@@ -41,22 +37,14 @@ fun CourseLengthChip(
                     color = colorResource(R.color.background_border),
                     shape = RoundedCornerShape(4.dp),
                 ).padding(horizontal = 10.dp, vertical = 4.dp),
-        verticalAlignment = Alignment.CenterVertically,
     ) {
-        Image(
-            painter = painterResource(R.drawable.icon_length),
-            contentDescription = null,
-        )
-
-        Spacer(modifier = Modifier.width(6.dp))
-
         Text(
             text =
-                formattedMeter(
-                    context = LocalContext.current,
-                    meter = length.meter,
+                stringResource(
+                    R.string.course_item_distance_to_course_format,
+                    formattedMeter(LocalContext.current, distance.meter),
                 ),
-            color = colorResource(R.color.item_primary),
+            color = colorResource(R.color.point_primary),
             fontSize = 14.sp,
             fontWeight = FontWeight.Bold,
         )
@@ -65,6 +53,6 @@ fun CourseLengthChip(
 
 @PreviewLightDark
 @Composable
-private fun CourseLengthChipPreview() {
-    CourseLengthChip(Length(5000))
+private fun CourseDistanceChipPreview() {
+    CourseDistanceChip(Distance(5))
 }
