+++ android/app/src/main/java/io/coursepick/coursepick/presentation/customcourse/component/CourseNavigationButton.kt
@@ -1,4 +1,4 @@
-package io.coursepick.coursepick.presentation.component
+package io.coursepick.coursepick.presentation.customcourse.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
